fix: Start to clean up namespaces.
