+++ Assets/Plugins/Source/Editor/Platforms/Linux/LinuxBuilder.cs
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-namespace PlayEveryWare.EpicOnlineServices.Build
+namespace PlayEveryWare.EpicOnlineServices.Editor.Build
 {
     using UnityEditor;
