+++ Assets/Plugins/Source/Editor/Platforms/macOS/MacOSBuilder.cs
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-namespace PlayEveryWare.EpicOnlineServices.Build
+namespace PlayEveryWare.EpicOnlineServices.Editor.Build
 {
     using System.IO;
     using UnityEditor;
