+++ Assets/Plugins/Source/Editor/Utility/EACUtility.cs
@@ -20,7 +20,7 @@
 * SOFTWARE.
 */
 
-namespace PlayEveryWare.EpicOnlineServices.Build
+namespace PlayEveryWare.EpicOnlineServices.Editor.Build
 {
     using UnityEditor.Build.Reporting;
     using UnityEditor;
