+++ Assets/Tests/PlayMode/ConfigTests.cs
@@ -22,12 +22,13 @@
 
 using NUnit.Framework;
 
-namespace PlayEveryWare.EpicOnlineServices.Tests
+namespace PlayEveryWare.EpicOnlineServices.Tests.Config
 {
     using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Reflection;
+    using Config = EpicOnlineServices.Config;
 
     public class ConfigTests
     {
