+++ Assets/Tests/PlayMode/EOSSessionsManagerTests.cs
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-namespace PlayEveryWare.EpicOnlineServices.Tests.IntegrationTests
+namespace PlayEveryWare.EpicOnlineServices.Tests.Services.Sessions
 {
     using Epic.OnlineServices;
     using Epic.OnlineServices.Sessions;
