+++ Assets/Tests/PlayMode/Extensions/CustomMappedEnumTestBase.cs
@@ -20,7 +20,7 @@
  * SOFTWARE.
  */
 
-namespace PlayEveryWare.EpicOnlineServices.Tests
+namespace PlayEveryWare.EpicOnlineServices.Tests.Extensions
 {
     using NUnit.Framework;
