+++ com.playeveryware.eos/Runtime/Core/Extensions/AuthScopeFlagsExtensions.cs.meta
