+++ com.playeveryware.eos/Runtime/Core/Extensions/FileInfoExtensions.cs
