+++ com.playeveryware.eos/Runtime/Core/Extensions/IntegratedPlatformManagementFlagsExtensions.cs
