+++ com.playeveryware.eos/Runtime/Core/Extensions/ListExtensions.cs
