+++ com.playeveryware.eos/Runtime/Core/Extensions/PlatformFlagsExtensions.cs
